# Implemented CountBy (Issue 104)