--- a/MoreLinq.Test/CountByTest.cs
+++ b/MoreLinq.Test/CountByTest.cs
@@ -99,5 +99,11 @@ namespace MoreLinq.Test
 
             countByKeys.AssertSequenceEqual(groupByKeys);
         }
+        
+        [Test]
+        public void CountByIsLazy()
+        {
+            new BreakingSequence<string>().CountBy(x => x.Length);
+        }
     }
 }