--- a/MoreLinq.Test/TestExtensions.cs
+++ b/MoreLinq.Test/TestExtensions.cs
@@ -73,7 +73,7 @@ namespace MoreLinq.Test
                 yield return split;
         }
 
-        internal static void Add<TKey, TValue>(this List<KeyValuePair<TKey, TValue>> list, TKey key, TValue value)
+        internal static void Add<TKey, TValue>(this IList<KeyValuePair<TKey, TValue>> list, TKey key, TValue value)
         {
             list.Add(new KeyValuePair<TKey, TValue>(key, value));
         }