--- a/MoreLinq/project.json
+++ b/MoreLinq/project.json
@@ -6,9 +6,9 @@
   "authors": [
     "MoreLINQ Developers."
   ],
-  "copyright": "\u00a9 2008 Jonathan Skeet. Portions \u00a9 2009 Atif Aziz, Chris Ammerman, Konrad Rudolph. Portions \u00a9 2010 Johannes Rudolph, Leopold Bushkin. Portions \u00a9 2015 Felipe Sateler, \u201csholland\u201d. Portions \u00a9 Microsoft. All rights reserved.",
+  "copyright": "\u00a9 2008 Jonathan Skeet. Portions \u00a9 2009 Atif Aziz, Chris Ammerman, Konrad Rudolph. Portions \u00a9 2010 Johannes Rudolph, Leopold Bushkin. Portions \u00a9 2015 Felipe Sateler, \u201csholland\u201d. Portions \u00a9 2016 Leandro F. Vieira (leandromoh). Portions \u00a9 Microsoft. All rights reserved.",
   "packOptions": {
-    "releaseNotes": "Adds new operators: Assert, AtLeast, Cartesian, EndsWith, EquiZip, Exclude, FallbackIfEmpty, FullGroupJoin, GroupAdjacent, Incremental, Interleave, Lag, Lead, NestedLoops, OrderBy, PartialSort, PartialSortBy, Partition, Permutations, Random, RandomDouble, RandomSubset, Rank, RankBy, Repeat, RunLengthEncode, Segment, Slice, SortedMerge, StartsWith, Subsets, TagFirstLast, ThenBy, TraverseBreadthFirst, TraverseDepthFirst, Windowed, ZipShortest. See also https://github.com/morelinq/MoreLINQ/wiki/API-Changes.",
+    "releaseNotes": "Adds new operators: Assert, AtLeast, Cartesian, CountBy, EndsWith, EquiZip, Exclude, FallbackIfEmpty, FullGroupJoin, GroupAdjacent, Incremental, Interleave, Lag, Lead, NestedLoops, OrderBy, PartialSort, PartialSortBy, Partition, Permutations, Random, RandomDouble, RandomSubset, Rank, RankBy, Repeat, RunLengthEncode, Segment, Slice, SortedMerge, StartsWith, Subsets, TagFirstLast, ThenBy, TraverseBreadthFirst, TraverseDepthFirst, Windowed, ZipShortest. See also https://github.com/morelinq/MoreLINQ/wiki/API-Changes.",
     "summary": "This project enhances LINQ to Objects with extra methods, in a manner which keeps to the spirit of LINQ.",
     "owners": [
       "Jon Skeet",