@atifaziz 2016-11-01
It would be nice to attempt a version that, like `GroupBy`, maintains the order of the keys with respect to the source sequence. If it turns out too complicated then we can revert but it seems worth a try.