@atifaziz 2016-11-02
> Could you please rebase this over master? It looks like you added this on top of the [2.0 beta 2](https://github.com/morelinq/MoreLINQ/releases/tag/2.0-beta-2), which has a different project structure.

Ignore this part. I was looking at a stale branch that was probably rewritten. Sorry.