@atifaziz 2016-11-03
I've take a crack at implementing the [optimizations I suggested](https://github.com/morelinq/MoreLINQ/pull/207#discussion_r86244247). The tests are passing but feel free to review & let me know if you think we may need to add some tests to for corner cases where the optimization may fail.

## Benchmarks

I measured the overall benefit of the optimizations. Here's the benchmark code (using [BenchmarkDotNet](https://www.nuget.org/packages/BenchmarkDotNet/0.9.9)) that runs a `CountBy` on an ordered and unordered sequence of 10K integers (with duplicates) as keys:

```
using System;
using System.Linq;
using BenchmarkDotNet.Attributes;
using MoreLinq;

public class Benchmark
{
    static readonly int[] UnorderedData;
    static readonly int[] OrderedData;

    static Benchmark()
    {
        var xs = Enumerable.Range(1, 1000)
                           .SelectMany(x => Enumerable.Repeat(x, 10))
                           .ToArray();
        OrderedData = xs;
        UnorderedData = xs.OrderBy(_ => Guid.NewGuid()).ToArray();
    }

    [Benchmark]
    public static void CountByOnOrderedKeys() =>
        OrderedData.CountBy(e => e).Consume();

    [Benchmark]
    public static void CountByOnUnorderedKeys() =>
        UnorderedData.CountBy(e => e).Consume();
}
```

### Benchmark Environment

```
Host Process Environment Information:
BenchmarkDotNet.Core=v0.9.9.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-6600U CPU 2.60GHz, ProcessorCount=4
Frequency=2742185 ticks, Resolution=364.6727 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=32-bit RELEASE
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1080.0

Type=Benchmark  Mode=Throughput  
```

### Results without Optimizations

                 Method |      Median |     StdDev |
----------------------- |------------ |----------- |
   CountByOnOrderedKeys | 478.9549 us | 45.6884 us |
 CountByOnUnorderedKeys | 481.1692 us | 29.9369 us |

### Results with Optimizations

Method |      Median |     StdDev |
----------------------- |------------ |----------- |
   CountByOnOrderedKeys | 210.2153 us |  3.8459 us |
 CountByOnUnorderedKeys | 334.2708 us | 17.2886 us |

### Conclusion

The optimizations have improved the speed of `CountBy`! 🙌 In the unordered case (`CountByOnUnorderedKeys`), it runs 30% faster due to fewer dictionary look-ups. In the ordered case (`CountByOnOrderedKeys`), the improvement is 2 folds by avoiding look-ups as long as the key is not changing.