@leandromoh 2016-11-05
@atifaziz you've done a great job with the optimization! I would begin it today but you were faster lol 
I have only one doubt: why did you use 

```
dic.Comparer.GetHashCode(prevKey) == dic.Comparer.GetHashCode(key) 
&& dic.Comparer.Equals(prevKey, key)
```
instead of just `dic.Comparer.Equals(prevKey, key)` ?