@atifaziz 2016-11-05
@leandromoh It's not worth calling `Equals` for performance reasons if the hash code is already different because that's what a dictionary would do internally too. Comparing hash codes can be faster in many cases. For example, `Equals` for a string will check for every character but a hash code comparison is just an integer equality check and therefore much faster (same goes for structural value types that are compounds of primitives). If hash codes match then a full equality comparison of the actual content is needed because the hash is just a digest & can be the same for several values of a type.

Since you're happy with the optimisations then I can merge this PR if you can just fix [the type parameter reference doc issue](https://github.com/morelinq/MoreLINQ/pull/207#discussion_r86241035).