@leandromoh 2016-11-10
`Can we address this now as the last item? Should be done then. Thanks!`

Done!