@atifaziz 2016-11-11
Merged into `master` by a02b8f97af48670c2ff628e0d31bc6b878343064.

Thanks for adding this and working through all the reviews.

---
Note that I squashed your commits in the final merge except the optimizations because I felt it would be good to keep a change history of those in case anything needs to be reverted or reviewed if ever a bug is reported.