@atifaziz 2016-11-11
I know [I said MoreLINQ is in lock down for 2.0 release](https://github.com/morelinq/MoreLINQ/pull/197#issuecomment-255585332) but I'm being bad here and letting this one slip… 😇 

`CountBy` is now published in [2.0 beta 6](https://www.nuget.org/packages/morelinq/2.0.0-beta06).