@atifaziz 2016-11-12
[Documentation published](https://morelinq.github.io/2.0/ref/api/html/Overload_MoreLinq_MoreEnumerable_CountBy.htm).