@leandromoh 2016-11-12
@atifaziz that's a good new! When do you plan to publish the release 2.0 with the other methods?

I open the link to see CountBy documentation and perceive that my name is not in the footer with others contributor. When do you plan to add it ?