@atifaziz 2016-11-13
> When do you plan to publish the release 2.0 with the other methods?

Soon-ish.

> my name is not in the footer with others contributors.

Sorry, missed that one but fixed now by 58d6c62f1aacb4a53fa4faba2486516bacbd363e. BTW, that's an aggregate of copyright notices, not contributors list.